+++ e2e/cypress/integration/content.js
@@ -1,10 +1,8 @@
-import { 
-  ITEMS_PANELS_SELECTORS, 
-} from '../elements/panels-selectors'
-import { 
-  EDITABLE_FORM_SELECTORS, 
+import { ITEMS_PANELS_SELECTORS } from '../elements/panels-selectors'
+import {
+  EDITABLE_FORM_SELECTORS,
   FORM_SELECTORS,
-  USER_FORM_SELECTORS
+  USER_FORM_SELECTORS,
 } from '../elements/form-types-selectors'
 import { NOTIFICATION_SELECTOR } from '../elements/notification-selectors'
 import { ACTION_SELECTORS } from '../elements/actions-selectors'
@@ -21,12 +19,13 @@ describe('test content', function () {
       .should('contain', 'Users')
   })
   it('creates a folder as Admin, then deletes it', function () {
-
     // Create Folder
     cy.addContent('Test folder', 'test-folder', 'btnAddFolder')
 
     // Delete Folder
-    cy.get(`[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-folder']`).within(() => {
+    cy.get(
+      `[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-folder']`
+    ).within(() => {
       cy.get(ACTION_SELECTORS.delete).click()
     })
     cy.get(ACTION_SELECTORS.btnConfirmModal).click()
@@ -34,15 +33,17 @@ describe('test content', function () {
   })
 
   it('creates an item as Admin, modifies it and delete it', function () {
-
     // Create Item
-    cy.intercept(`/db/container/test-item/@canido?permissions=guillotina.AddContent,guillotina.ModifyContent,guillotina.ViewContent,guillotina.DeleteContent,guillotina.AccessContent,guillotina.SeePermissions,guillotina.ChangePermissions,guillotina.MoveContent,guillotina.DuplicateContent,guillotina.ReadConfiguration,guillotina.RegisterConfigurations,guillotina.WriteConfiguration,guillotina.ManageAddons,guillotina.swagger.View`).as('canido')
+    cy.intercept(
+      `/db/container/test-item/@canido?permissions=guillotina.AddContent,guillotina.ModifyContent,guillotina.ViewContent,guillotina.DeleteContent,guillotina.AccessContent,guillotina.SeePermissions,guillotina.ChangePermissions,guillotina.MoveContent,guillotina.DuplicateContent,guillotina.ReadConfiguration,guillotina.RegisterConfigurations,guillotina.WriteConfiguration,guillotina.ManageAddons,guillotina.swagger.View`
+    ).as('canido')
 
     cy.addContent('Test Item', 'test-item', 'btnAddItem')
 
-
     // Modify Item
-    cy.get(`[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-item']`).click()
+    cy.get(
+      `[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-item']`
+    ).click()
     cy.wait('@canido')
     cy.get(
       `[data-test='${EDITABLE_FORM_SELECTORS.prefixEditableField}-title']`
@@ -53,12 +54,14 @@ describe('test content', function () {
       cy.get(EDITABLE_FORM_SELECTORS.field).clear().type('Test Modified Item')
       cy.get(EDITABLE_FORM_SELECTORS.btnSave).click()
     })
-    
+
     cy.get(NOTIFICATION_SELECTOR).should('contain', 'Field title, updated!')
     cy.get(`[data-test='${BREADCRUMB_SELECTORS.prefixItem}-container']`).click()
 
     // Delete Item
-    cy.get(`[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-item']`).within(() => {
+    cy.get(
+      `[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-item']`
+    ).within(() => {
       cy.get(ACTION_SELECTORS.delete).click()
     })
     cy.get(ACTION_SELECTORS.btnConfirmModal).click()
@@ -80,7 +83,9 @@ describe('test content', function () {
     cy.get(NOTIFICATION_SELECTOR).should('contain', 'Content created!')
 
     // Modify Item
-    cy.get(`[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-user']`).click()
+    cy.get(
+      `[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-user']`
+    ).click()
     cy.get(USER_FORM_SELECTORS.username).type('Test Modified User', {
       force: true,
     })
@@ -90,7 +95,9 @@ describe('test content', function () {
     cy.get(`[data-test='${BREADCRUMB_SELECTORS.prefixItem}-users']`).click()
 
     // Delete Item
-    cy.get(`[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-user']`).within(() => {
+    cy.get(
+      `[data-test='${ITEMS_PANELS_SELECTORS.prefixItem}-test-user']`
+    ).within(() => {
       cy.get(ACTION_SELECTORS.delete).click()
     })
     cy.get(ACTION_SELECTORS.btnConfirmModal).click()