+++ e2e/example/components/guillotina.js
@@ -3,7 +3,7 @@ import { Guillotina } from 'react-gmi'
 import { Auth } from 'react-gmi'
 import { Login } from 'react-gmi'
 import { getClient } from 'react-gmi'
-import { ClientProvider , RequiredFieldsForm} from 'react-gmi'
+import { ClientProvider, RequiredFieldsForm, Icon, TdLink } from 'react-gmi'
 
 import '../node_modules/react-gmi/dist/css/style.css'
 
@@ -15,8 +15,7 @@ const registry = {
   // to register views around guillotina objects paths
   paths: {},
   // default views for content types
-  views: {
-  },
+  views: {},
   // forms for adding content, they are fired
   // throught an action
   forms: {
@@ -29,7 +28,66 @@ const registry = {
   },
   behaviors: {},
   components: {},
-};
+  itemsColumn: {
+    Container: () => {
+      const smallcss = { width: 25 }
+      const mediumcss = { width: 120 }
+
+      return [
+        {
+          label: '',
+          child: (m) => <td style={smallcss}>{<Icon icon={m.icon} />}</td>,
+        },
+        {
+          label: 'type',
+          child: (m) => (
+            <TdLink style={smallcss} model={m}>
+              <span className="tag">{m.type}</span>
+            </TdLink>
+          ),
+        },
+        {
+          label: 'id/name',
+          child: (m, navigate, search) => (
+            <TdLink model={m}>
+              {m.name}
+              {search && (
+                <React.Fragment>
+                  <br />
+                  <span className="is-size-7 tag is-light">{m.path}</span>
+                </React.Fragment>
+              )}
+            </TdLink>
+          ),
+        },
+        {
+          label: 'created',
+          child: (m) => (
+            <td style={mediumcss} className="is-size-7 is-vcentered">
+              {m.created}
+            </td>
+          ),
+        },
+        {
+          label: 'depth',
+          child: (m) => (
+            <td style={mediumcss} className="is-size-7 is-vcentered">
+              {m.item.depth}
+            </td>
+          ),
+        },
+        {
+          label: 'modified',
+          child: (m) => (
+            <td style={mediumcss} className="is-size-7 is-vcentered">
+              {m.updated}
+            </td>
+          ),
+        },
+      ]
+    },
+  },
+}
 
 export default function App() {
   const [isLogged, setLogged] = React.useState(auth.isLogged)
@@ -43,7 +101,7 @@ export default function App() {
 
   return (
     <ClientProvider client={client}>
-      {isLogged && <Guillotina auth={auth} url={url} registry={registry}/>}
+      {isLogged && <Guillotina auth={auth} url={url} registry={registry} />}
       {!isLogged && (
         <div className="columns is-centered">
           <div className="columns is-half">