+++ src/guillo-gmi/lib/utils.js
@@ -1,3 +1,5 @@
+import { stringToSlug } from './helpers'
+
 export const formatDate = (str) => {
   const d = new Date(str)
   const minutes = d.getMinutes() < 10 ? `0${d.getMinutes()}` : d.getMinutes()
@@ -22,7 +24,7 @@ export function getNewId(id = '') {
   const suffix = '-copy-'
   const rgx = new RegExp(`($|${suffix}\\d*)`)
 
-  return id.replace(rgx, (r) => {
+  return stringToSlug(id).replace(rgx, (r) => {
     const num = parseInt(r.replace(suffix, '') || '0')
     return `${suffix}${num + 1}`
   })